refactor: create search package
