+++ engine/language-server/src/test/scala/org/enso/languageserver/search/Suggestions.scala
@@ -1,4 +1,4 @@
-package org.enso.languageserver.runtime
+package org.enso.languageserver.search
 
 import java.util.UUID
